Make directoy before file creaion
